[e2e] Test username annotation lifecycle
This commit updates the e2e tests to validate BYOH node annotations.
Criteria was added to the waitForNodes to ensure that the BYOH annotation and
the username annotation are present, decipherable, and correct.

Also, a reconfiguration test case was added to ensure nodes are properly
updated when the SSH private key is changed to a random key.