[wsvalidator][cm_controller] Validate Services CM
This commit provides a library function to parse and validate a services
ConfigMap. The validation is around the depenencies and priorities of
services defined within the CM in order to ensure bootstrap services
always start before controller services according to WICD's expected schema.

It also adds logic to the configmap_controller to delete + recreate
a valid services ConfigMap if an invalid one exists already when reconciling.