[cm_controller] Bump MaxConcurrentReconciles
This commit changes the ConfigMap controller's max concurrent reconciles to 2
(from default of 1) now that there are 2 distinct respnsibilities for the
ConfigMap controller. If the ConfigMap controller of WMCO is busy configuring
BYOH instances, the controller will not be able to react to events regarding
the services ConfigMap, until it is no longer busy. There could be a period of
time in which the services ConfigMap could have incorrect contents, which
could then be picked up by WICD running on various instances. To reduce the
risk of this, the MaxConcurrentReconciles option can be provided when
initalizing the ConfigMap controller. This will enable the controller to
process service ConfigMap events that occur while WMCO is handling its
BYOH configuration responsibilities.