[wicd] Introduce cleanup command
This commit implements the WICD cleanup command, which stops and removes
all services on the instance that are defined in the services.
The cleanup command will delete the associated Node object by default.
An optional `--preserveNode` flag can be supplied to deconfigure a node
in-place, meaning the instance's Node object will not be deleted.

The functionality includes Windows service dependency handling to stop and
delete any dependents when removing a service, regardless of if the dependents
are described in the services ConfigMap.