[manager] Tweak DeleteService logic
Makes the DeleteService method upon the Windows service manager ensure a service
is stopped before deleting it. Behavior was also changed to no longer fail when
trying to delete a non-existant service. This prevents unsafe deletions from
having to be handled at each call site, as was needed before this change.