[wicd] Introduce cleanup command
This commit implements the WICD cleanup command, which stops and removes
all services on the instance that are defined in the services.
The cleanup command will delete the associated Node object by default.
An optional `--preserveNode` flag can be supplied to deconfigure without
deleting the instance's Node object, which is useful during node upgrades.

This commit also implements a best-effort cleanup strategy when the expected
version of the services CM cannot be found (e.g. node object or desired version
annotation not found). Best effort cleanup is a combination of removing:
- any Windows services with the OpenShift managed tag in its description
- all services in the latest-created services ConfigMap, if one exists

The functionality includes Windows service dependency handling to stop and
delete any dependents when removing a service, regardless of if the dependents
are described in the services ConfigMap.