[ignition][windows] Extract & copy bootstrap files
This commit makes WMCO responsible for creating files needed to start kubelet:
bootstrap kubeconfig, kubelet config, kubelet CA certificate, cloud-config file
These files are extracted from the ignition contents when the operator pod
starts, and are then copied from the payload to each instance when reconciling.

The worker ignition is now parsed and stored in a global variable on operator
start. This is to prevent havin to GET ignition each time a node is configured.