[nodeconfig] Extract & copy bootstrap files
This commit makes WMCO responsible for creating files needed to start kubelet:
- bootstrap kubeconfig
- kubelet config
- kubelet CA certificate
- cloud-config file (if enabled)

The bootstrap kubeconfig is created from the contents of the MCO node
bootstrapper secret. This change is needed since bootstrap kubeconfig is not
available in the MachineConfig object we now read for ignition spec. The
kubelet configuration is no longer created from a embedded file, and instead
through a more maintable object.
The other required files are extracted from the ignition contents.

The files are re-populated every time a node reconciles to ensure they are
created  using the latest ignition spec.These files are copied directly to
Windows instances when reconciling nodes.