[services_cm] Add proxy vars to services CM
This commit expands the services ConfigMap data schema to hold a new field, a
list of name/value pairs representing environment variables to be set on Windows
instances.

The expected services ConfigMap content is also expanded to hold the proxy
variables in environments with a cluster-wide proxy enabled, or empty otherwise
(values read from WMCO container environment, which is managed by OLM). If the
cluster-wide proxy variables change at any point, the ConfigMap is re-created to
hold the new values.