Make WMCO copy trusted CA bundle to instances
This commit copies over the CA bundle with all trusted cluster certs to Windows
instances. To ensure this file is always up to date, this is done, if needed,
during node cofiguration as well as when there are any events on the ConfigMap
holding the trusted CA (watched and managed by WMCO, with data injected by CNO).