[controller] WICD import of trusted certs
This commit makes WICD import all certs present in the CA bundle file into the
Windows instance's system store, if they are not already present.
If any certs needed to be imported, WICD will communicate to WMCO to initiate
a system reboot so that all node services can load the new certs when restarted.