[cleanup] WICD cleans up all managed services
The best effort cleanup of WICD is expanded to look at service descriptions
to cleanup all services that WICD started, even if they are not in the
current iteration of the Windows services ConfigMap.
This also prevents a user from being able to sneak in services into the CM
that WICD would never cleanup.